moving group_membership into its own file
